Update to support atm regional history output
Regional history output capability under SE dycore does not work properly. In typical format of use, it leads to invalid values and incorrect number of time frames in the history file. This update resolve the problems by addressing the issues related to the handling of the record dimension in the netcdf file for patch output. Note that the CAM tag has already addressed this problem.

  Modified:   components/cam/src/control/cam_history_support.F90
  Modified:   components/cam/src/utils/cam_pio_utils.F90

[BFB]