Remove -DUSE_PIO2 from FCINCLUDES to avoid assembler output with XL compilers
The XL Fortran compiler creates assembler (*.s) files rather than object (*.o)
files when -DUSE_PIO2 is present in the build options. Not all uses of -D<something>
seem to lead to this behavior: for example, -DUS and -DS lead to the generation
of assembler rather than object files, but -DfooS does not.

Since MPAS is essentially always built with the XL compilers with GEN_F90=true,
we only need -DUSE_PIO2 to be included in the definition of CPPFLAGS, and the include
paths only need to be added to FCINCLUDES.