Updated default GHG for CMIP6 transisent.
The last year in the GHG file for CMIP6 is 2014. I originally
extrapolated it to 2015.  However, the run would then die on the last
timestep in 2014, because a half-timestep in 2015 is needed to do
output, which requires concnetrations in 2016.  With this commit, the
default GHG file is updated to the new file that is extrapolated to
2016, and hence the model can run to the end of 2014.

[BFB, except for final timestep of 2014 for CMIP6 transient simulation]