COMPOSE: Rework names, transport/SL options; CAAS; super levels.
1. Rename the monolithic files as follows:
    ir.cpp -> slmm.cpp
    qlt.cpp -> cedr.cpp
Modify the compose_mod routine prefixes similarly. (Eventually we'll use the
COMPOSE library installation, but for now, integration is much easier with
monolithic files I create from the COMPOSE code base.)

2. Rework semi-Lagrangian transport options:

  ! Tracer transport algorithm type:
  !     0  spectral-element Eulerian      [default]
  !     1  classical semi-Lagrangian (SL)
  !     2  cell-integrated remap SL

  ! Constrained density reconstructor for SL property preservation; not used if
  ! transport_alg = 0:
  !     0  none
  !     1  Cobra
  !     2  QLT    [default]
  !     3  CAAS

  ! If true, check mass conservation and shape preservation. The second
  ! implicitly checks tracer consistency.
  logical, public  :: semi_lagrange_cdr_check = .false.

3. Add global CAAS as an option. Global CAAS is as simple as a CDR gets, so it's
a good to have as an MPI benchmark if nothing else.

4. Implement super levels. The idea is to run a CDR over 8 adjacent vertical
levels, a super level. (I could expose this number as an option, but for now
it's hardcoded in an enum to 8.) This has the immediate effect of reducing comm
volume by 8x. Later, it will also make adaptive QLT potentially have to do no
comm at all besides the adaptation handshake of one scalar per tree edge.