Fix a PGI Fortran if short circuit bug on Titan
On Titan, the if short circuit is valid for the PGI Fortran, in a
debug mode with an option 'Ktrap=fp', however, it is invalid. This bug
caused a floating point error from the clm 'SurfaceResistanceMod.F90'
in the debug test of 'SMS_D_Ln5.ne4_ne4.FC5AV1C-L', because the values of
'soilp_col' is NaNs in the if statement at line 157.

Fixes: issue #2104

[BFB] - Bit-For-Bit