Removes physics_update_intr call and renamed it to physics_update
This commit automates addfld and outfld calls and some other
processes. ptend names are updated to reflect the processes
and they are all unique now. It also adds a capability whereby
we can add new variables for outfld calls if we want.

See confluence for a more detailed description about these tags.