Moves CNP balance check initialization after dynamic area change
The initialization of CNP balance check is now moved after changes
due to dynamic PFTs have been accounted for in various CNP pools.
This change is a precursor for supporting dynamic column area,
which will break conservation at column level. Additional CNP balance
check will be added in the future for quantities that are not conserved
at column level but are conserved at grid level.