Add sfcMassBalApplied fields and use for AMs
A nontrivial amount of SMB typically occurs at locations that are not
identified as having ice by the cellMask at the end of the time step.
This makes using the global/regional stats for SMB not very useful for
assessing global mass balance terms.  This commit remedies this by
introducing a field called sfcMassBalApplied that contains the SMB
values actually applied at every grid cell within the apply_mass_balance
routine in the advection module.  This includes negative SMB where the
SMB is sufficient to completely remove all ice from a grid cell, and in
that situation it accounts for the fact that the actual ice removed is
likely to be less than the full ablation value specified in the
sfcMassBal field.  Thus, sfcMassBal can be thought of as a “potential”
SMB, and sfcMassBalApplied is what is actually used.  When  the supplied
sfcMassBal field is expected to be used exactly and conservation is
required (say in E3SM), additional work would be needed to track the
difference.

A second field called groundedSfcMassBalApplied is also added that is
analogous but just for grounded ice.  The difference of the two yields
the applied SMB over floating ice.  These values are required for the
global/regional stats calculations for SMB over just grounded or
floating ice. (Trying to apportion the sfcMassBalApplied field to
grounded and floating ice at the end of the time step is impossible.)

These new fields are applied to the global and regional stats fields
relating to SMB.  I tested to confirm that the global and regional
values are identical when the region is the global domain.  I also
confirmed that totalGroundedSfcMassBal is the same as groundingLineFlux
for a simple test case run out to steady state.  Prior to these changes
that was NOT the case - the difference was about 15%.  The test
configuration led to a noisy steady state groundingLineFlux, but
averaging over about 250 time steps yielded agreement between the two
variables to the 4th decimal place.

With these changes, these two fields can be used to assess mass balance
for a grounded ice sheet, at least at steady state.  For transient
situations, there still is a missing term related to the grounded ice
volume associated with grounding line advance or retreat.