Add field and stat to track volume ungrounded on a time step
After advection occurs, this routine compares the grounding and floating
masks for the state prior to advection and the state afterwards.  Cells
the changed from grounded to floating or marked as a positive change and
negative for vice versa in the new groundedToFloatingThickness field.

The calculation for flux across the grounding line is moved to AFTER the
new calculation for change from grounded to floating so that it is valid
for the new grounding line.

The groundedToFloatingThickness field is used by global stats members to
calculate total values.