Fix GL migration flux
The implementation of GL migration flux previous commit was generally
correct, but failed to close the grounded mass balance in some
situations.  This was because the SMB/BMB was being applied after the GL
migration flux was calculated, and the SMB/BMB could alter where the ice
is grounded or floating.

This commit moves the calculation of the GL migration flux, and also the
GL flux, to *after* the BMB/SMB is applied so that the grounded and
floating extents are at their final locations for the time step.  This
commit also simplifies the implementation a bit.