COMPOSE: Speed up calc_sphere_to_ref; silence Kokkos warnings.
ALE_departure_from_gll does not normalize the output point to the sphere. This
is mathematically OK for intersection and departure-cell-determination
calculations, and even for ref <-> sphere calculations. But it's an efficiency
loss for the final one because calc_sphere_to_ref in slmm.cpp was iterating too
many times because of the norm(q,2) != 1. This commit speeds up that calculation
by almost 2x.

Thanks to Oksana for asking about this.

Oksana also pointed out the Kokkos warnings issue; warnings get printed on every
rank. Fixed in this commit.