Use topounits as basis for tropical tree area in fire module
The area of two tropical tree types was being referenced through
fractional weight of columns on gridcells. This was switched to use
a fractional weight of columns on topounits to allow same fire module
behavior with and without multiple topographic units.
The results are BFB with max_topounits=1, and roundoff different with
max_topounits>1.

[BFB] - Bit-For-Bit
L2-178