Remove 'varsso' field from MPAS-Atmosphere initialization
The 'varsso' field is not needed by the current MPAS-Atmosphere code, and
in fact it is never even read by the model itself, so we can safely remove
this field from the MPAS-Atmosphere initialization core.

Note: although the varsso field was listed in the Registry.xml file under
a heading indicating that this field is used by the GWDO scheme, this same
field is used in WRF for the 'topo_wind' option, and not the GWDO scheme.