use shr_reprosum_calc in seq_diag_avect_mct
When BFBFLAG is set to true and when INFO_DBUG > 1, the
routine seq_diag_avect_mct uses a reproducible sum algorithm
that is not as accurate as the algorihm implemented in
shr_reprosum_calc. In particular, when summing a vector of INFs,
the current algorithm returns zero. Here we replace the
existing algorithm with a call to shr_reprosum_calc.

This change is BFB for standard usage (INFO_DBUG == 1).
It is not BFB with respect to the associated diagnostic,
written to cpl.log, when INFO_DBUG > 1. However, these diagnostics are
not used in the simulation, and simulation results will BFB.

[BFB]