Add routine to remove small grounded islands
This routine finds and eliminates very small islands that lead to
unrealistic velocities in the Albany velocity solver.  Specifically,
this finds one- and two-cell masses of ice that are surrounded by open
ocean and eliminates them by sending them to the calving flux.  Grounded
such ice masses would be islands, floating would be iceberges (which we
don't want anyway).  The routine is applied after each of the calving
options, but not applied if no calving routine is applied.