Output MPI task to compute node mapping
Currently MPI task to compute node mapping information is
output in two locations, once in CAM, where it is
truncated after the first 256 MPI tasks, and once in CLM,
where it is truncated after the first 100 MPI tasks,
both only for the respective component. This is not useful in current
production runs. The use of environment variables, such as
MPICH_CPUMASK_DISPLAY on Cray systems, generate data that are
unnecessarily verbose for our needs. Here a share routine is
introduced that writes out one line per compute node. Each line
contains the compute node name and the list of MPI tasks assigned
to that node for a given communicator. This is then called
in the driver and writes out the task-to-node mapping for the
entire coupled model. A separate branch will then introduce
this into the individual components, replacing the current logic
in both CAM and CLM, for example.

The share routine also optionally returns the number of compute
nodes and the task-to-node mapping, which is needed in the
internal CAM load balancing.

[BFB]