Output CAM MPI task to compute node mapping using share routine
MPI task to compute node mapping information in CAM
is output one line per task, and only for the first 256 MPI
tasks. A new share routine in CIME (shr_taskmap_write) is available
that writes out one line per compute node. Each line
contains the compute node name and the list of MPI tasks assigned
to that node for a given communicator. This share routine also
optionally returns the number of compute nodes and the
task-to-node mapping, information that is needed in the
internal CAM load balancing. Here the CAM logic in spmdinit (in
spmd_utils.F90) is removed and a call to shr_taskmap_write in
atm_init_mct (in atm_comp_mct.F90) is added, recording information for
all MPI tasks and in a more compact representation for each CAM
instance.

[BFB]