Output CLM MPI task to compute node mapping using share routine
MPI task to compute node mapping information in CLM
is output one line per task, and only for the first 100 MPI
tasks. A new share routine in CIME (shr_taskmap_write) is available
that writes out one line per compute node. Each line
contains the compute node name and the list of MPI tasks assigned
to that node for a given communicator. Here the CLM logic in spmd_init
(in spmdMod.F90) is removed and a call to shr_taskmap_write in
lnd_init_mct (in lnd_comp_mct.F90) is added, recording information for
all MPI tasks and in a more compact representation for each CLM
instance.

[BFB]