Decrease cost of assign_chunks
The routine assign_chunks is called as part of create_chunks in
phys_grid_init when determining what decomposition to use in the
CAM physics. It attempts to assign chunks to processes in such
a way as to minimize the amount of MPI communication required in
d_p_coupling and p_d_coupling when moving data between the dynamics
and physics decompositions. It uses a greedy heuristic that
identifies which process in the dynamics decomposition was assigned
the most columns in a given chunk, and assigns the chunk to that
process for the physics decomposition. If this process has already
been assigned its target number of chunks, the algorithm then attempts
to assign the chunk to the process assigned the second-most columns.
etc.

As implemented currently, there is a doubly nested loop of sizes
nprocs and nchunks, respectively, when using phys_loadbalance=2.
For ne120 experiments on 86400 processes with 4 threads per process
(for example, on Cori-KNL at NERSC), the number of iterations is
86400*86400*4, and the computation takes over 7 minutes. Here the
algorithm is reorganized, decreasing the loop over the number of
processes to something less than
  min(max(number of columns in a chunk), nprocs).
For this example, the number of iterations decreases to 3*86400*3 ,
and the execution time decreases correspondingly.

Note that at process count scale, especially when using
phys_loadbalance=2 (which uses a wrap map), identifying the
process with the maximum number of columns may not be worth the
cost. Finding any process that was assigned one of the columns in the
chunk is probably the best that can be achieved. If the new algorithm
still proves too expensive for these extreme cases, we can then look
to implement a hybrid algorithm, using the new algorithm when the
number of processes is moderate or when not using phys_loadbalance=2,
and something less costly when the number of processes is large.
However, experiments with the new algorithm show the cost to be less
than one second on Cori-KNL even for the extreme example described
previously.

Fixes #2496

[BFB]