Migrate air pressure to topounit data structure
Following the same pattern used for air temperature, bottom layer air pressure
from the atmosphere is defined as a forcing variable at the topographic unit
level. The sbetr and clm_interface components have not been altered in this
commit - additional work will be required after the rest of the forcing variables
are migrated.

[BFB] - Bit-For-Bit
L2-178