Revert "Automatic code formatting commit"
This reverts commit 90c77d73c53b1a42dbff034bed793c65ad6dbf3f.

Revert an accidental automated format commit and push to master.

We should discuss this part of the dev-ops setup before using it. In addition,
there are at least two formatting conventions currently used that I believe are
highly nonstandard: no space between a control keyword and open parens, and
putting the pointer annotation on the variable rather than the type.