Reorganizing the directory tree, plus some cmake work
Grouped all source stuff in the src folder, and all physics parametrizations
in the src/physics folder.

Also, added a cmake folder, with som kokkos and compiler flags scripts in it.
The compiler flags script is mostly copied from Hommexx, with the difference
that the concept of horix/vert openmp is removed, and we kept only the
gnu and intel compilers. We should add other compilers if/when they are needed.