Added a small structure to hold time stamps
The time stamp is not the same as std::time, in that it does not refer to
the OS time. It is instead the time that the simulation refers to.

The class is not modifiable, so that one cannot tamper with time stamps.
At some point, someone must be able to modify time stamps. When
you have identified what class/function should have that privilege, come
back and grant them friendship.

Notice that there's no point in implementing getters. The whole point of
time stamps is to be able to compare them. We provide (friend) functions
to compare two time stamps, which should be the only points of
'interaction' that 'normal' customers of TimeStamp will need.