Adding some pieces to the driver.
The driver, after creating and init-ing the processes, now queries them for
all the required/computed fields, and adds them to the field repo (PBUF).
At this point, the atm processes do not own any "field", they only have
headers (with metadata).

Then, when everything is registered, and fields are init-ed (i.e., properly
allocated), it passes back a field to the processes. At this point, the
field *should* contain a valid kokkos view. All fields handled by the repo
are 1d, and the atmosphere processes are responsible of reshaping them
to better suit their needs (obviously, without changing global size of
the view).