COMPOSE: Make global CAAS #rank-invariant.
Previously, CEDR's global CAAS implementation always used MPI_Allreduce. Now it
can take a user-provided all-to-all reducer. In the case of Homme, this custom
reducer is a wrapper to repro_sum. The result is that global CAAS now gives
results invariant to MPI problem distribution. (QLT already does because of its
tree.)