Adjust fast ice routine to apply to grounded ice
Experience has shown that really fast velocities can occur on grounded
islands of very small size, which make up an unrealistic geometry.  It
would be best to eliminate these locations, but a backup alternative is
to extend the fast ice routine to apply to both floating and grounded
ice.  That is done here.