Adjust calculation of GL depth
Before I was using the GL mask in cellMask, which was not working
correctly.  Even if it was, I realized that that mask include grounded
cells adjacent to open ocean, which we do NOT want to include.  This
modified method averages the depths at cells that are floating but
adjacent to grounded cells.