Remove check against MPI_TAG_UB
The `mpi_attr_get` function, which is used to determine the value of
MPI_TAG_UB, is not available in the mpi-serial implementation. This
function was being used to throw an exception when the max MPI tag was
exceeded and report what the max tag is, however, MPI itself will throw
this exception (but not report what the max tag is), so it is not
necesary. Removing it fixes mpi-serial builds.

Fixes #2890