COMPOSE: In ISL MPI compute comm round 2, skip empty messages.
In the compute phase of ISL MPI, there are two comm rounds. In round 1, ranks
trade requests for data at departure points. In round 2, tracer data are traded
to fulfill the requests. This commit implements the following: If in round 1,
the departure point list for a rank is empty, then the corresponding tracer data
message is skipped in round 2.