Added a 'group' atmosphere process.
This class allows a group of processes to be seen as a single process from
outside. This helps the interface with the driver, which does not need
to know how the processes are organized internally.

The AtmosphereProcessGroup is where the logic for Parallel and Sequential
splitting has to be handled, making the splitting invisible from outside,
effectively hiding the implementation detail.