Major changed in the Field structure
The Field class is now only capable of storing 1d views. Consequently,
the first template argument of Field is no longer a multidimensional
array data type, but it is just the scalar type. This change is
mainly suggested by the fact that we were not accessing the field's
values from within Field. Therefore, there was no need for having
the DataType encoded in the class, nor it had any added value.
One can still extract a N-dimensional view from the field, using
the templated method 'get_reshaped_view', provided that the requested
data type is compatible with the allocation. That is, the requested
data type must have no compile-time dimensions, or, if it does,
it must have compile time dimensions compatible with the allocation
size, which may be different from the dimensions stored in the
FieldIdentifier. This can happen in case there is 'padding', to
accommodate customers that want to use the data by reinterpreting
as Pack<T,N>.