Fixing the allocation properties logic.
We no longer require that the value types of two different requests have
sizes that are one a multiple of the other. We simply make sure that the
allocation can accommodate both. E.g., say we have a 1d field of 9 Real's.
Request 1 wants Pack<Real,4>, and request 2 wants Pack<Real,5> (for some
reason). We compute the the number of packs n_i of each request, and store the
maximum of n_i*s_i, where s_i is the size of the i-th request.