Update the constant CO2 override value to be CMIP6-compliant
In BCRC and BDRC compsets, the constant value of CO2 seen by the
atmosphere radiation code is specified using the user namelist
override variable, co2vmr_rad.  This variable was still set to the
CMIP5 value for PI CO2 concentrations.  This PR updates the value to
the CMIP6 number, which is consistent with the default constant CO2
values specified elsewhere in the code, and provided to land and ocean
BGC components in the BGC experiment compsets.

[Non-BFB]
[NML]

See confluence for a more detailed description about these tags.