Implement info_taskmap_comp support in CAM
A recent PR added the env_run.xml parameter INFO_TASKMAP_COMP
and the associated cime_pes namelist parameter info_taskmap_comp
to enable (> 0) or disable (<= 0) the output of the
task-to-node mapping for the component models. If enabled,
it further specifies whether to use a compact (== 1) or
a more verbose (> 1) output format. Here the necessary control
logic is added for CAM.

[BFB]