Replace "merge" with if statement
Replace the merge statement with tau in the denominator with an if
statement to protect against divide by zero. For some reason, the "if
true" source for the merge statement was being evaluated before checking
the conditional with the PGI compiler, which threw a floating point
exception. Replacing the merge statement with an if block protects
against this.