Avoid divide-by-zero in zb computation for Jablonowski & Williamson initialization
The initialization routine for the Jablonowski and Williamson test case could
previously generate a divide by zero in the computatation of zb and zb3 due to
areaCell(nCells+1) being zero and appearing in the denominator of a term.

One way to avoid this divide by zero would have been to simply set
areaCell(nCells+1) to some reasonable non-zero value. However, other test cases
that also compute zb and zb3 added an if-test to compute zb and zb3 only for
edges that border an owned cell. For uniformity, the same approach is used for
the Jablonowski and Williamson test case. Since the value of the zb and zb3
field only matters for owned edges (in the test case setup), there is no change
to results.