Remove a pointless test.
Remove test_p3_main. It was used early in development to keep things
stable. It's superseded by test_p3_ic.

test_p3_main gives an FPE for a trivial reason that is essentially b/c this was
a build test, not a run one. test_p3_ic does not have an FPE b/c it gets the
proper data to run p3_main.