Add ocean/ice shelf coupling to components
This commit adds a first cut at applying the ocean/ice shelf coupling
fields to the OCN and GLC components.

It is a squash of work done primarily by Jeremy Fyke between 2015 and 2017, with
updates by Jon Wolfe to make it compatible with changes to E3SM
since then. It is BFB for existing compsets.

This commit introduces the component changes, and requires the previous
commit that adds the CIME changes

[BFB] - Bit-For-Bit