Homme: Reorganized the design of RemapFunctor
I stripped what was model specific (i.e., required knowledge on the state),
and isolated it in a separate class. I also gave (hopefully) more explanatory
names to the class _RemapFunctorRSplit.

In the current design, RemapFunctor stores a RemapStateAndThicknessProvider
object, which provides src/tgt thicknesses as well as the states to remap
(the latter only for rsplit=0). This class stores a RemapStateProvider
object, which provides the states to remap. This class is (will be) different
in the theta model implementation, due to a change in the state structure,
but also due to the fact that theta remaps 2 variables that are not part of
the state, but are instead 'increments' of 2 states (what you would get
with diff(x) in matlab, to give an idea).

This last difference has more consequences: in the fortran implementation
of the theta model, these differences need to be computed from the states,
then the remap happens on these differences, and then the states are updated
with the remapped differences. The first and third step did not exist in
the remap functor as it was implemented for preqx, so the RemapStateProvider
also offers two methods, to pre and post process the states. For preqx,
these two methods do nothing, while for theta they will perform the calculation
of the increments, as well as the states update.