Fix bug in create_chunks
In the routine create_chunks in the file
components/cam/src/physics/cam/phys_grd.F90
is logic to determine the number of columns assigned to each virtual
SMP during the dynamics. As part of this a mapping from each column to
the owner SMP is computed using the physics (lon-lat) ordering of the
columns. To count the number of columns per SMP, this mapping array is
then accessed as if the entries had been assigned using the native
dynamics ordering of the columns. This error however does not prevent
the correct calculation of the number of columns per SMP, and the
column-to-SMP mapping array is then deleted, so the inconsistency is
not repeated and does not introduce an error.

However, when enhancing the load balancing logic to
use data on the computational cost of columns, this
column-to-smp mapping is needed. To introduce clarity in the
current logic and in support of this future enhancement,
the ordering used is changed to that of the natural dynamics
ordering of the columns (not necessarily lon-lat), and used
consistently throughout create_chunks.

[BFB]