Homme: introduced a buffers manager for functors
The buffers are now allocated in one shot by the FunctorsBuffersManager (FBM),
and each functors then has a Buffers structure (privately defined), which
simply subviews the FBM buffers with functor-specific names. This is already
somewhat useful in a couple of places, where Caar and ESF where grabbing
unused buffers whose name had nothing to do with them. However, this will
become even more important with the theta model. In fact, without the FBM,
we would either have one ElementsBuffers per model, or use, say, a buffer
called temperature_virt to store non-hydrostatic pressure, or something like
that. This way, each functor can use verbose names. We may even reduce
the number of total buffers needed by a couple.

The idea is the following:
1) create functors
2) the functors request buffers
3) the FBM allocates, based on all the requests received
4) the functors grab the memory from the FBM

All the requests must be submitted to the FBM *before* the allocation happens.