Remove EAM dependencies from standalone P3 modules
This commit removes all dependencies on EAM modules for the
standalone p3 modules; micro_p3.F90 and micro_p3_utils.F90.

The only dependency that is retained is for the designation of real
and integer precision (r8 and i8) which is taken from CIME in EAM.
A compile time directive is used to determine if running in standalone
mode or in EAM.  If in EAM then the CIME defintion is used for consistency.

In addition, the real type and integer type variables have been changed
from r8 and i8 to rtype and itype respectively.  This is to make the
variable names more general and remove the implied implementation (8-byte real)
that the old names have.