Added new aqua-planet compsets
I opted to use a different naming convention to get away from the "FC5AV1C" in favor of "F-EAMv1-AQP1". The long name of the compsets then looks like:
2000_EAM%V1-AQUA_SLND_SICE_DOCN%AQP1_SROF_SGLC_SWAV
It seems that we eventually want to move away from using "CAM5" in our compsets, since that's a differnet model, so this is a small push in that direction. I also needed to add a new use case, which ensures that the constant droplet number ind prescribed ozone options are used by default. I also set the new use case "aquaplanet_EAMv1" to use the solar data file "atm/cam/solar/ape_solar_ave_tsi_1365.nc" to be consistent with how the CAM6 aqua-planet compsets are configured. I did minimal analysis of the resulting test simulations, but I have verified that the topography is flat and everything is symetric about the equator, which is a good improvement on the old compsets that had mountains of water. A new default cami file will be added in the next commit.

 Changes to be committed:
	new file:   components/cam/bld/namelist_files/use_cases/aquaplanet_EAMv1.xml
	modified:   components/cam/cime_config/config_component.xml
	modified:   components/cam/cime_config/config_compsets.xml
	new file:   components/cam/cime_config/usermods_dirs/aquap/user_nl_cpl