Add x2o direction to tbot and dztbot
This forces the attribute vectors to be updated by the time that
prep_glc_calculate_subshelf_boundary_fluxes() is called.  Without this,
Sg_tbot and Sg_dztbot were still 0 when that routine was executed,
leading to a divide by zero.

There may be a subtler way to force the update, but this has the added
benefit of allowing these fields to be visualized on the ocean grid in
the coupler history file for debugging purposes.