Ensure that ierr value from jedi_da package setup is not overwritten
The logic in atm_setup_packages for setting up the 'jedi_da' package will
set ierr to a non-zero value if an error occurs. However, further down in
the atm_setup_packages routine, the call to atmphys_setup_packages overwrites
the ierr value. In cases where the jedi_da package setup fails but physics
package setup encounters no errors, the model will not stop as intended.

This commit assigns the return value of atmphys_setup_packages to a local
variable, and increments the ierr return value if the local variable is
non-zero, ensuring that the model will stop if errors are encountered in
either the setup of the jedi_da package or of any physics packages.