Add measurement and output per chunk of atmospheric physics cost
Measurement of cost on a per chunk basis allows for a more direct
evaluation of physics load balancing. This will be important as new
physics capabilities are added that require new physics load balancing
schemes, for example, superparameterization and elevation
classes. Here this capability is added. A new logical namelist
parameter, phys_chnk_cost_write, is added to enable the output of
these data to a file named, for the moment, atm_chunk_costs.txt. (In
the future, the name of this file may be set via the namelist.) The
output is disabled by default, but the data are always collected,
and the collection overhead is very low. The namelist parameter can be
set to true in the user_nl_cam file. This can be added to env_run.xml
in the future if desired.

[BFB]