Add option to load balance atm. physics wrt column cost estimates
For both superparameterization and elevation class schemes the
current load balancing options addressing diurnal and seasonal cycles
and land/ocean differences may not address the major atmospheric
physics load imbalances. One approach being investigated is to provide
as input computational cost estimates for each column, and then
assign columns to chunks to approximately load-balance the chunks
with respect to these cost estimates. Here this capability is
introduced, but if all columns cost estimates are identical (default
is a value of 1.0), then the current algorithm is used. Using
column cost estimates is implemented in the same logical structure
as the current algorithm, being restricted to load balancing within
a virtual SMP with the option of pairing each column assignment with
its latitude-longitude "twin". Support for supplying column cost
estimates is limited to the SE dycore first, and even this is just a
dummy entry as the data must then be provided by the SE dycore via an
optional parameter in the get_horiz_grid_d routine.

[BFB]