Moved atmosphere_process_group to share
It does not belong in control. Also, using the generic util::Factory class,
and eliminating the atmosphere_process_factory.*pp files.

Each atmosphere process header must define a creator function (see e.g. the
surface_coupling.hpp file). This will have to be registered in the process
factory before any AtmosphereProcessGroup is created, since that class relies
on the factory for automatic (and agnostic) process creation.

More work to come on this.