Removed dependency of control on dynamics.
This stems from the fact that, when relying on homme, dynamics needs some
configure time parameters in order to build as a library. This forces
dynamics to be created just-in-time for new executables. On the other hand,
control would like to implement some tests, and would need to link against
dynamics. To solve this, we make control dynamics-agnostics. Control should
only bother about processes organization (init, ordering, run, clean-up),
fields book-keeping (through a field repository), and handling surface coupling.
In this context, there's no need to know about how dymamics works.
Note: the driver can still check the type of each atm process, so it can
      still check, for instance, that there is a dynamics process in the
      processes dag.