Improvements to init_atmosphere core to enable regional static interpolation
This commit adds three new fields to the Registry.xml file,
bdyMask{Cell,Edge,Vertex}, plus new code to ensure that pixels of static
data that are mapped to a cell actually belong to that cell. This latter change
is needed for limited-area meshes, since the remapping of source pixels for
terrain, land use, soil category, monthly albedo, monthly vegetation fraction,
and maximum snow albedo using a path-based search does not guarantee that
the pixel actually lies in the cell where the search terminates (in particular,
in the case of pixels outside of a limited-area mesh, where the search
terminates at the boundary cell closest to, but not containing, the pixel).

If the bdyMask{Cell,Edge,Vertex} fields are not present in the input file, they
will receive a default value of zero, which will effectively deactivate the new
logic added in this commit, giving results as in prior versions of the code.

At present, a bdyMaskCell value of 7 (via the parameter nBdyLayers) is assumed
to represent the value for cells on the outermost edge of a limited-area mesh;
however, if regional meshes make use of different numbers of relaxation cell
layers or specified cell layers, this will need to be generalized.

Eventually, some of the new functions in this commit should be considered
for migration to mpas_geometry_utils.F.