bugfix introduced in PR #66
This commit fixes two bugs introduced in PR #66.

1) Moving the cloud fraction calculation into the micro_p3_interface layer
introduced a bug that could lead to floating point invalids.  This was
because the full pcols array was passed to the cloud fraction routine which
expected an array with its:ite as the first dimension.  This has been fixed
by passing cldm arrays into the cloud fraction routine as icldm(its:ite,kts:kte)
and so on.

2) Similarly, moving cloud fraction to the interface layer introduced an issue
with generating baselines in the scream code since icldm, lcldm and rcldm were
shifted from OUT to IN for p3_main but were not properly initialized in the C++
code.  This has been fixed by initializing the arrays.

3) Lastly, code clean up, ast was an INput to p3_main, but calculating the cloud
fraction in the interface layer made it unecessay to pass ast to p3_main.  This
has been removed as a variable passed to p3_main.