PIO library detection cleanup for PIO version check
Previously, if the user specified `USE_PIO2=true`, the Makefile would find the
PIO library that was associated with PIO version 2. Likewise, if `USE_PIO2=true
` was not specified.

This creates a problem when we check for a PIO version. So, regardless of what
version the user specified, find *any* PIO version library installation
location. That way, any version of PIO can be reliably deteced, and the user
can be alerted that they have or have not specified the correct version.