Introduced an abstract remapper class
This class is responsible to remap fields from one layout to another. Probably,
the two most common cases will be remapping physics to/from dynamics, and
remapping due to different mpi distributions (this will happen only if/when
we enable concurrent atmosphere processes, aka parallel splitting).

The concrete implementations need to be able to

1) detect the layout fields, and mark it as 'source' or 'target'
2) detect if two layouts (src and tgt) are 'compatible', meaning whether
   it is possible to remap from one to the other
3) remap in either direction (from src to tgt layout and viceversa)

For performance reasons, we set the source and target fields once and
for all, so that the remapper can do any needed setup once. E.g., the
physics-dynamics remap will need some halo exchange. Setting this up
has some costs, so we want to do it only once, rather than at every
call to remap.

Note: source and target here are just place-holder names. The remapper
      must be able to perform the remap in either direction.