Removed pointless cmake variables
SCREAM_LIBRARY_DIRS are pointless, since cmake already knows the directory
where to find the targets that it builds. Linking against, say, scream_share,
does not require to specify its directory, since cmake already knows it.
A proof of this is that the value that was stored in this variable was
${CMAKE_BINARY_DIR}/src, which contained no libraries.

As a consequence, remove UNIT_TEST_EXTRA_LIBRARY_DIRS from CreateUnitTest,
since its purpose was to provide the directory of the scream dynamics library
(which is not known a priori, but only upon specification of homme's config
parameters). Since cmake already knows where to find the dynamics library,
this variable is pointless.