Avoid invalid values when remapping the MODIS maximum snow albedo field
Mismatches in the land mask between the MPAS mesh and the implied land mask
in the MODIS maximum snow albedo field have been found to lead to MPAS grid
cells for which we have not mapped any source values. As a quick fix, simply
set the maximum snow albedo field in MPAS for such grid cells to the fill value.

It's not clear why this is an issue for the maximum snow albedo field but not
for the MODIS monthly albedo field.

In future, a better solution that finds nearby, non-missing maximum snow
albedo values to assign to MPAS grid cells would be preferable.