Added the concept of 'grid'.
A grid is simply a map that stores information about columns. In particular,
a grid consists of a certain number of columns; for each column, we store
the (local) id of the element owning it, the i-j coordinates within the
element, and the global id of the column.

A grid must provide:
 - its type: Undefined, Dynamics, Physics, or Radiation
 - the number of columns owned
 - the elem-i-j-gid map (see above)

We provide a default implementation for a grid, templated on the type.