Storing a grid pointer inside the field identifier.
The pointer can be set at construction time (if available) or later. Setting
the grid after construction is only allowed if the stored grid is of type
Undefined, which signifies that a default "null"-like grid was stored.
If the stored grid is not of type 'Undefined', an exception is thrown.