Change in atmosphere process interface
The communicator must now be passed at construction time (as opposed as
during the call to initialize). This changes the type specification of
the atmosphere factory as well.

This will allow to init homme during the construction of the homme dynamics
atmosphere process, allowing to be able to have the identifiers of all the
required/computed fields upon construction completion. This will allow the
driver to let the field repo build all the fields and then set them in
the atmosphere processes *before* calling the initialize method.