Updates to allow running pg2 with topography
The new routine fv_phys_to_dyn_topo() called from read_inidat(), takes the topography on the pg2 grid and copies it over to the dynamics grid, which includes the boundary exchange step.

A note to anyone trying to generate a new topography file: I found that mapping a np4 topo to file to the pg2 grid after it had been smoothed by HOMME to be a good solution for getting an adequately smoothed topo field. I experimented with a "nearest neighbor" smoothing method, but was unable to get this too work, although this was likely due to a bug in the code, so using a smoothing routine outside of HOMME will probably work.

	modified:   components/cam/src/dynamics/se/fv_physics_coupling_mod.F90
	modified:   components/cam/src/dynamics/se/inidat.F90