HOMME: added Session struct with an option on whether to terminate or throw
This static variable allows an external library using Hommexx to prevent
Hommexx from terminating execution when an error occurs. If the option
m_throw_instead_of_abort is true, instead of terminating execution
(via call to finalize_hommexx_session and MPI_Abort), Hommexx
throws an exception. The host application is then free to decide what to
do with it.

Note: if the host application wants to set the option to false, it must do
      so *after* hommexx session has been inited. In fact, the function
      initialize_hommexx_session initializes the option to true.