Replace hard-coded field list with generated list
In prep_glc_mod, there were two places where a list of coupling fields
from OCN were hardcoded in calls to seq_map_map.  This commit replaces
those hard-coded lists with a variable that is generated in seq_flds_mod
from the definitions of the coupling fields that are required.